fix: address second-round audit review feedback
- Reject SetQueuePaused values other than 0/1 with InvalidInstructionData,
  and assign the validated flag directly (with a regression test).
- Re-apply the single `aligned` definition in add_item_with_id (the C17
  hunk lost while resolving the merge with the C15 restructuring).
- Standardize logging on msg! and drop the log(String) helper, which only
  added a format! allocation on top of msg!.
- Make QUEUE_TTL_SECONDS a u64 (it is a duration, not a timestamp) and
  drop the max(0) clamps at both use sites.
- Rename the shadowed `acc` slice to `body` in QueueAccount::load.

+++ program/tests/test.rs
@@ -299,7 +299,7 @@ async fn run_test() {
     let current_slot = banks.get_sysvar::<Clock>().await.unwrap().slot;
     context.warp_to_slot(current_slot + 1_000).unwrap();
     let mut clock = banks.get_sysvar::<Clock>().await.unwrap();
-    clock.unix_timestamp = clock.epoch_start_timestamp + QUEUE_TTL_SECONDS + 1_000;
+    clock.unix_timestamp = clock.epoch_start_timestamp + QUEUE_TTL_SECONDS as i64 + 1_000;
     context.set_sysvar(&clock);
 
     // Purge expired requests
